Move TerraHandler to handlers sub-module.
