+++ terra/VteObject.py
@@ -27,7 +27,7 @@ from gi.repository import Gtk, Vte, GLib, Gdk, GdkX11, GObject
 import terra.terra_utils as terra_utils
 from terra.interfaces.preferences import Preferences
 from terra.config import ConfigManager
-from terra.handler import TerraHandler
+from terra.handlers import TerraHandler
 from terra.i18n import t
 from terra.interfaces.prog_dialog import ProgDialog
 from terra.interfaces.win_dialog import WinDialog
