+++ terra/handlers/__init__.py
@@ -1,0 +1,6 @@
+"""
+This sub-module contains various handlers for terra.
+"""
+
+
+from terra.handlers.terra_handler import TerraHandler
