+++ terra/interfaces/preferences.py
@@ -24,7 +24,7 @@ import shutil
 from gi.repository import Gtk, Gdk, GdkPixbuf, GdkX11
 
 from terra.config import ConfigManager
-from terra.handler import TerraHandler
+from terra.handlers import TerraHandler
 from terra.i18n import t
